# server/cmd/api-public/docs/swagger.yaml
basePath: /
definitions:
...
  book.ImagesResponse:
    properties:
      colors:
        items:
          type: string
        type: array
      inkUrl:
        type: string
      mainUrl:
        type: string
      mediumUrl:
        type: string
      pixelBase64:
        type: string
      smallUrl:
        type: string
    required:
    - colors
    - inkUrl
    - mainUrl
    - mediumUrl
    - pixelBase64
    - smallUrl
    type: object
  book.MetadataResponse:
    properties:
      ageGroups:
        items:
          type: string
        type: array
      authors:
        items:
          type: string
        type: array
      description:
        type: string
      ean:
        type: string
      genres:
        items:
          type: string
        type: array
      isbn:
        type: string
      moods:
        items:
          type: string
        type: array
      pacingTypes:
        items:
          type: string
        type: array
      pageCount:
        type: integer
      publishedAt:
        type: string
      publisher:
        type: string
      settings:
        items:
          type: string
        type: array
      themes:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - authors
    - description
    - isbn
    - pageCount
    - publishedAt
    - publisher
    - title
    type: object
  book.Response:
    properties:
      archivedAt:
        type: string
      createdAt:
        type: string
      id:
        type: string
      images:
        $ref: '#/definitions/book.ImagesResponse'
      metadata:
        $ref: '#/definitions/book.MetadataResponse'
      updatedAt:
        type: string
    required:
    - createdAt
    - id
    - images
    - metadata
    - updatedAt
    type: object
  controller.HTTPError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Public API for @codeserk / media
  title: Media API / Public
  version: "1.0"
paths:
...
  /api/v1/books/isbn/{isbn}:
    get:
      consumes:
      - application/json
      description: Search books using a query
      operationId: searchBooks
      parameters:
      - description: ISBN to look for
        in: path
        name: isbn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/book.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.HTTPError'
      summary: Search
      tags:
      - books
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  Signature:
    in: header
    name: x-signature
    type: apiKey
  SignatureApp:
    in: header
    name: x-signature-app
    type: apiKey
swagger: "2.0"